feat: separate component
