+++ src/scenes/components/TimepointExplorer/TimepointViewerActions.hooks.tsx
@@ -2,8 +2,7 @@ import { useCallback, useMemo } from 'react';
 import { IconName } from '@grafana/data';
 import { trackTimepointViewerActionClicked } from 'features/tracking/timepointExplorerEvents';
 
-import { CheckType } from 'types';
-import { getCheckType, getExploreUrl } from 'utils';
+import { getExploreUrl } from 'utils';
 import { useLogsDS } from 'hooks/useLogsDS';
 import { useMetricsDS } from 'hooks/useMetricsDS';
 import { useSceneVarProbes } from 'scenes/Common/useSceneVarProbes';
@@ -11,7 +10,6 @@ import { useTimepointExplorerContext } from 'scenes/components/TimepointExplorer
 import { StatelessTimepoint } from 'scenes/components/TimepointExplorer/TimepointExplorer.types';
 import { getIsInTheFuture } from 'scenes/components/TimepointExplorer/TimepointExplorer.utils';
 import { getProbeNameToUse } from 'scenes/components/TimepointExplorer/TimepointViewerActions.utils';
-import { useFaroSessionLink } from 'scenes/components/TimepointExplorer/TimepointViewerFaroSession.hooks';
 
 interface Action {
   icon: IconName;
@@ -21,21 +19,12 @@ interface Action {
   href?: string;
 }
 
-export type FaroActionStatus = 'loading' | 'no-session' | 'available';
-
-export interface FaroAction {
-  status: FaroActionStatus;
-  href?: string;
-  tooltip: string;
-  onClick?: () => void;
-}
-
 export function useTimepointViewerActions(timepoint: StatelessTimepoint) {
   const logsDS = useLogsDS();
   const metricsDS = useMetricsDS();
   const { check, checkType, currentAdjustedTime, handleViewerStateChange, listLogsMap, viewerState, timepoints } =
     useTimepointExplorerContext();
-  const [_, viewerProbeName, viewerExecutionIndex] = viewerState;
+  const [, viewerProbeName] = viewerState;
   const logsQuery = { expr: `{job="${check.job}", instance="${check.target}", probe="${viewerProbeName}"} | logfmt` };
   const metricsQuery = {
     expr: `{job="${check.job}", instance="${check.target}", probe="${viewerProbeName}"}[$__range]`,
@@ -56,23 +45,6 @@ export function useTimepointViewerActions(timepoint: StatelessTimepoint) {
   const exploreMetricsURL = getExploreUrl(metricsDS?.uid!, [metricsQuery], {
     from: timepoint.adjustedTime,
     to: timepoint.adjustedTime + timepoint.timepointDuration,
-  });
-
-  const isBrowserCheck = getCheckType(check.settings) === CheckType.Browser;
-  const selectedExecution =
-    viewerProbeName !== undefined && viewerExecutionIndex !== undefined
-      ? listLogsMap[timepoint.adjustedTime]?.probeResults?.[viewerProbeName]?.[viewerExecutionIndex]
-      : undefined;
-  const executionId = selectedExecution?.labels.execution_id;
-  const {
-    data: faroSession,
-    isLoading: isFaroLoading,
-    isFetched: isFaroFetched,
-  } = useFaroSessionLink({
-    executionId: executionId ?? '',
-    from: timepoint.adjustedTime,
-    to: timepoint.adjustedTime + timepoint.timepointDuration + timepoint.config.frequency,
-    enabled: isBrowserCheck && Boolean(executionId),
   });
 
   const handlePreviousTimepoint = useCallback(() => {
@@ -98,48 +70,6 @@ export function useTimepointViewerActions(timepoint: StatelessTimepoint) {
       });
     }
   }, [checkType, nextTimepoint, listLogsMap, probeVar, handleViewerStateChange]);
-
-  const faroAction = useMemo<FaroAction | null>(() => {
-    if (!isBrowserCheck) {
-      // Not a browser check — Frontend Observability doesn't apply here.
-      return null;
-    }
-
-    if (!executionId) {
-      // Browser check, but no execution selected yet. Keep the action visible so it
-      // doesn't pop in/out of the header, just degrade it until there's something to check.
-      return {
-        status: 'no-session',
-        tooltip: 'Select an execution to check for a Frontend Observability session',
-      };
-    }
-
-    if (isFaroLoading || !isFaroFetched) {
-      return {
-        status: 'loading',
-        tooltip: 'Checking for a Frontend Observability session…',
-      };
-    }
-
-    if (!faroSession?.href) {
-      return {
-        status: 'no-session',
-        tooltip: 'No Frontend Observability session was found for this execution',
-      };
-    }
-
-    return {
-      status: 'available',
-      href: faroSession.href,
-      tooltip: 'View Frontend Observability session',
-      onClick: () => {
-        trackTimepointViewerActionClicked({
-          checkType,
-          action: 'view-frontend-observability-session',
-        });
-      },
-    };
-  }, [isBrowserCheck, executionId, isFaroLoading, isFaroFetched, faroSession?.href, checkType]);
 
   const actions = useMemo<Action[]>(() => {
     const actions: Action[] = [
@@ -190,5 +120,5 @@ export function useTimepointViewerActions(timepoint: StatelessTimepoint) {
     handlePreviousTimepoint,
   ]);
 
-  return { actions, faroAction };
+  return { actions };
 }
